Moved volunteers into list component
