+++ src/contexts/VolunteerListContext.js
@@ -1,0 +1,3 @@
+import { createContext } from "react"
+
+export const VolunteerListContext = createContext();
